+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH6K3N7ASCSTRT2EQRXRY5SK"
+        "01KH6W3Y8421F0DY7DZF5F448T"
     )
     r"""The unique identifier of the data source"""